projeto_5/chatgpt/requirements.txt
streamlit
openai>=1.0.0
python-dotenv
PyMuPDF